\name{SSDforR}
\alias{SSDforR}

\title{
List of all functions in SSD for R
}
\description{
Lists all the functions available in the SSD for R package.
}
\usage{
SSDforR()
}


\references{

Auerbach, Charles, and Zeitlin Wendy. SSD for R: An R Package for Analyzing Single-Subject Data. Oxford University Press, 2014. p18

Go to www.ssdanalysis.com for more information.
}
\author{Charles Auerbach, PhD & Wendy Zeitlin, PhD; Yeshiva University,  Wurzweiler school of social work

}

\examples{SSDforR()
}

