\name{SSDforR}
\alias{SSDforR}

\title{
List of all functions in SSD for R
}
\description{
Lists all the functions available in the SSD for R package.
}
\usage{
SSDforR()
}


\references{
Go to www.ssdanalysis.com for more information.
}
\author{Charles Auerbach, PhD & Wendy Zeitlin Schudrich, PhD; Yeshiva University,  Wurzweiler school of social work

}

\examples{SSDforR()
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
