\name{SSDforR}
\alias{SSDforR}

\title{
SSD for R Visual and statistical analysis of single-subject-data
}
\description{
To list all function type SSDforR() in the console Visual analysis includes simple graphs, band graphs, SPC charts and more. Text can be added lines separating phases can be added. Statistical tests include descriptive, CDC, t-test, effect size, PND, PMD G-index, chi-square and more)
}
\usage{
SSDforR()
}


\references{
see www.ssdanalysis.com for more information
}
\author{Charles Auerbach & Wendy Zeitlin 
Wurzweiler school of social work

}

\examples{SSDforR()
}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
